Update systemc to 2.3.2
* Switch to cmake-based build of the systemc library.

Other fixes:
* De-ambiguify call to `sc_start(...)` which was required by new systemc version
* Removed custom build flag handling (we just pass on the `-std=XYZ`
  version to cmake)

+++ systemsim/stage2_virtual_hardware.h
@@ -44,7 +44,7 @@ public:
     mutexfile spike_receive_file; ///<filehandle to record received events on the wafer
     mutexfile spike_transm_file;  ///<filehandle to record transmitted events on the wafer
 
-	long _duration_in_NS;  //!< duration of the systemc simulation in nano-seconds, needed for the progress_bar.
+	int _duration_in_NS;  //!< duration of the systemc simulation in nano-seconds, needed for the progress_bar.
 
 	std::vector< std::unique_ptr<pcb> > pcb_i; //!< array containing all pcb instances
